refactor: folder path absolute & index path
