+++ frontend/apps/web/app/_components/index.ts
@@ -1,0 +1,1 @@
+export { default as Aside } from './Aside';
