+++ frontend/apps/web/components/client/Drawer/Drawer.tsx
@@ -1,9 +1,9 @@
 'use client';
 
+import { MODAL, useModal } from '@/hooks';
 import * as stylex from '@stylexjs/stylex';
 import { HTMLAttributes } from 'react';
 import { createPortal } from 'react-dom';
-import { MODAL, useModal } from '../../../hooks/useModal';
 
 interface DrawerProps extends HTMLAttributes<HTMLElement> {
   opened?: boolean;
@@ -11,7 +11,7 @@ interface DrawerProps extends HTMLAttributes<HTMLElement> {
   backgroundScroll?: boolean;
 }
 
-export const Drawer = ({
+const Drawer = ({
   opened = true,
   backgroundScroll = false,
   onClose,
@@ -50,6 +50,8 @@ export const Drawer = ({
       )
     : null;
 };
+
+export default Drawer;
 
 const fadeIn = stylex.keyframes({
   from: { opacity: 0 },
