+++ frontend/apps/web/components/client/Map/index.ts
@@ -1,0 +1,2 @@
+export { default as Map } from './Map';
+export { default as MapOverlay } from './MapOverlay';
