+++ frontend/apps/web/components/client/index.ts
@@ -1,0 +1,2 @@
+export * from './Drawer';
+export * from './Map';
