+++ frontend/apps/web/constants/index.ts
@@ -1,0 +1,1 @@
+export const DARK = '@media (prefers-color-scheme: dark)';
