+++ frontend/apps/web/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useModal, MODAL } from './useModal';
