+++ frontend/apps/web/stores/index.ts
@@ -1,0 +1,1 @@
+export * from './MapStore';
